+++ src/Extension/DefaultAttributes/ApplyDefaultAttributesProcessor.php
@@ -27,16 +27,12 @@ final class ApplyDefaultAttributesProcessor implements ConfigurationAwareInterfa
         /** @var array<string, array<string, mixed>> $map */
         $map = $this->config->get('default_attributes');
 
-        $document = $event->getDocument();
-        $walker   = $document->walker();
-        while ($event = $walker->next()) {
-            $node = $event->getNode();
+        // Don't bother iterating if no default attributes are configured
+        if (! $map) {
+            return;
+        }
 
-            // Only modify nodes when we first encounter them
-            if (! $event->isEntering()) {
-                continue;
-            }
-
+        foreach ($event->getDocument()->iterator() as $node) {
             // Check to see if any default attributes were defined
             if (($attributesToApply = $map[\get_class($node)] ?? []) === []) {
                 continue;