+++ src/Extension/Footnote/Event/GatherFootnotesListener.php
@@ -19,6 +19,7 @@ use League\CommonMark\Extension\Footnote\Node\Footnote;
 use League\CommonMark\Extension\Footnote\Node\FootnoteBackref;
 use League\CommonMark\Extension\Footnote\Node\FootnoteContainer;
 use League\CommonMark\Node\Block\Document;
+use League\CommonMark\Node\NodeIterator;
 use League\CommonMark\Reference\Reference;
 use League\Config\ConfigurationAwareInterface;
 use League\Config\ConfigurationInterface;
@@ -29,16 +30,10 @@ final class GatherFootnotesListener implements ConfigurationAwareInterface
 
     public function onDocumentParsed(DocumentParsedEvent $event): void
     {
-        $document = $event->getDocument();
-        $walker   = $document->walker();
-
+        $document  = $event->getDocument();
         $footnotes = [];
-        while ($event = $walker->next()) {
-            if (! $event->isEntering()) {
-                continue;
-            }
 
-            $node = $event->getNode();
+        foreach ($document->iterator(NodeIterator::FLAG_BLOCKS_ONLY) as $node) {
             if (! $node instanceof Footnote) {
                 continue;
             }